PS-7291: Allow option value validation to be run for 'set persist_only'
Currently there is no new value check done in case variable value
is updated via 'set persist_only'. It's written to config 'as is'.
This was done in order to disable checks for readonly variables
which are not applicable at runtime, otherwise it's not possible to
do 'set persist_only' for a readonly variable.
But as a result of this change some invalid value for a variable may
be written to a config file and crash the server after restart.
New value check can be enabled in this case to avoid such issues.
This check will differ a bit for usual sys variables and for
plugin variables.

For usual sys variables there is a basic check and also custom on_check
method provided. We could skip custom on_check only in case of
'set persist_only'. This would allow to do default checks for a value
(min/max value, allowed values) depending on its type and skipping some
specific checks which may be not applicable to runtime check for
readonly variables.

For plugin variables there is a default method which validates value
according to parameters provided at variable definition time. It is
also possible to provide custom validation method in which case it
will replace default one. Here again checks not applicable to readonly
variables at runtime will be done in custom validator. We still
could use default validation method with 'set persist_only' to
do default checks for provided value.

Add separate method which converts a variable value for suitable
string representation to be used with 'set persist_only'.