Avoiding unnecessary reseeks in primary key iterator
Upstream commit ID : fb-mysql-5.6.35/aa09ab287c1f34a3ad21e87627e3ecd5fb175b21
PS-8216 : Merge percona-202108

Summary:
Doing full index scans is currently more expensive than it needs to be with partial indexes. The problem is that for every distinct group, we have to reseek our primary key iterator. However, in most cases, a reseek is not necessary, as the primary key iterator is already at the correct position.

The change here is to track in `m_iterator_pk_position` the current position of the iterator relative to the current prefix. When seeking to the next prefix, we check if we are already positioned correctly, and then skip the seek if possible.

Also, as a minor optimization, use std::vector instead of std::map. This makes allocating/deallocating memory much cheaper.

Reviewed By: george-reynya

Differential Revision: D30681736

fbshipit-source-id: 4de4e0196d8