[upstream] PS-8772: Stall of instance after INSERT (...) SELECT.
https://jira.percona.com/browse/PS-8772

Some queries that used a big intrinsic temporary InnoDB table might have
led to InnoDB SE stall (especially when InnoDB buffer pool was small).
Also concurrent execution of those queries could have led to the same
problem or degraded performance.

The problem occured for queries that scanned through such intrinsic tables
using auto-created secondary key rather than clustered key.

What happened was that in such cases some of the pages belonging to
intrinsic table and accessed by the query stayed fixed in buffer pool
until the end of statement (and not until the end of operation on
specific row as it happens for normal InnoDB tables or only while they
are inspected as it happens for scans of intrinsic tables over clustered
key). As result these buffer fixed pages might have consumed the whole
buffer pool at some point and further statement progress was blocked.

Such behavior was result of combination of the following factors:
1) operations on intrinsic temporary tables are optimized and use single
   mini-transaction for all row operations in the statement (while
   for normal tables commit of mini-transaction happens after
   processing of each record.
2) normally, all buffer fixed pages are unfixed from memory only at
   commit of mini-transaction.
3) process of scanning over a key is optimized to immediately unfix pages
   once the cursor leaves them. However, when scan happens over a secondary
   key additional lookups on the clustered key are performed. And buffer
   pages which are fixed by these additional lookups stay fixed until
   the end of mini-transaction.

This fix solves the problem by ensuring that we commit mini-transaction
which is used for access to initrinsic temporary table after performing
operations on each 100 rows. This unfixes buffer pages touched by
mini-transaction making them available for further use by other or this
statements.

This patch might lead to 1-3% performance reduction in some of scenarios
which heavily use intrinsic temporary tables. OTOH we also have seen
significant performance improvements in other scenarios where big
intrinsic tables were used. So we consider this acceptable.

A better fix for this problem might be possible, but it is likely
to be more complex, and more appropriate for upstream in our
opinion.