PS-10183 [9.x]: Fix sporadic failures of rpl.rpl_insert_id test after mysql-9.5.0 merge.
This test has started to fail sporadically after mysql-9.5.0 made the
GTID mode default. The problem is caused by the fact that in case of
statements on MyISAM tables update of @@gtid_executed variable happens
before the tables are unlocked. As result there is a short window during
which INSERT statement is considered executed by code which checks if
replica caught up with source in include/rpl/sync_to_replica.inc script
(as it relies on GTID state if GTIDs are on), while results of the INSERT
are not yet visible to SELECT in other connections, provided that
concurrent inserts are allowed for MyISAM (@@concurrent_insert = 1,
this not the case on source, but is the case on replica in this test).

This is not a problem for case when we rely on binlog file name and
position for tracking replica state as those are updated after tables
are unlocked.

This commit solves the problem by disabling GTID mode for this test.